refactor: extract publish-devcontainer and ssh routes (Q2 Phase 2)
Move /sandboxes/{id}/publish-devcontainer and /sandboxes/{id}/ssh
routes to sandboxes-routes/features.route.ts.

sandboxes.route.ts: 2590 → 2112 lines (18% reduction)
