+++ apps/www/lib/routes/sandboxes.route.ts
@@ -1,5 +1,4 @@
 import {
-  getAccessTokenFromRequest,
   getUserFromRequest,
 } from "@/lib/utils/auth";
 import { getConvex, getConvexAdmin, type ConvexAdminClient } from "@/lib/utils/get-convex";
@@ -9,9 +8,8 @@ import { stackServerAppJs } from "@/lib/utils/stack";
 import { verifyTeamAccess } from "@/lib/utils/team-verification";
 import { env } from "@/lib/utils/www-env";
 import { api, internal } from "@cmux/convex/api";
-import type { Doc, Id } from "@cmux/convex/dataModel";
+import type { Id } from "@cmux/convex/dataModel";
 import { DEFAULT_MORPH_SNAPSHOT_ID } from "@/lib/utils/morph-defaults";
-import { RESERVED_CMUX_PORT_SET } from "@cmux/shared/utils/reserved-cmux-ports";
 import { parseGithubRepoUrl } from "@cmux/shared/utils/parse-github-repo-url";
 import { createRoute, OpenAPIHono, z } from "@hono/zod-openapi";
 import { HTTPException } from "hono/http-exception";
@@ -23,7 +21,6 @@ import {
   wrapPveLxcInstance,
 } from "@/lib/utils/sandbox-instance";
 import {
-  isPveLxcInstanceId,
   getInstanceById,
 } from "./sandboxes/provider-dispatch";
 import { getActiveSandboxProvider } from "@/lib/utils/sandbox-provider";
@@ -770,71 +767,7 @@ async function setupProviderAuth(
   return { providers: configuredProviders };
 }
 
-/**
- * Cmux instance metadata stored in Morph instance.metadata
- */
-interface CmuxInstanceMetadata {
-  app?: string;
-  userId?: string;
-  teamId?: string;
-}
-
-/**
- * Result of instance ownership verification
- */
-type VerifyInstanceOwnershipResult =
-  | { authorized: true; instanceId: string }
-  | { authorized: false; status: 403 | 404; message: string };
-
-/**
- * Verify that a user owns or has team access to a Morph instance.
- * Checks instance metadata for cmux app prefix and user/team ownership.
- */
-async function verifyInstanceOwnership(
-  morphClient: MorphCloudClient,
-  instanceId: string,
-  userId: string,
-  checkTeamMembership: () => Promise<{ teamId: string }[]>
-): Promise<VerifyInstanceOwnershipResult> {
-  let instance;
-  try {
-    instance = await morphClient.instances.get({ instanceId });
-  } catch {
-    return { authorized: false, status: 404, message: "Instance not found" };
-  }
-
-  const meta = instance.metadata as CmuxInstanceMetadata | undefined;
-
-  // Verify the instance belongs to cmux (accepts cmux, cmux-dev, cmux-preview, etc.)
-  if (!meta?.app?.startsWith("cmux")) {
-    return { authorized: false, status: 404, message: "Instance not found" };
-  }
-
-  // Check direct ownership
-  const isOwner = meta.userId === userId;
-  if (isOwner) {
-    return { authorized: true, instanceId };
-  }
-
-  // Check team membership if instance has a teamId
-  if (meta.teamId) {
-    try {
-      const memberships = await checkTeamMembership();
-      const isTeamMember = memberships.some((m) => m.teamId === meta.teamId);
-      if (isTeamMember) {
-        return { authorized: true, instanceId };
-      }
-    } catch {
-      // Failed to check team membership - continue to deny
-    }
-  }
-
-  return {
-    authorized: false,
-    status: 403,
-    message: "Forbidden - not authorized to access this instance",
-  };
-}
+// NOTE: CmuxInstanceMetadata, VerifyInstanceOwnershipResult, verifyInstanceOwnership moved to sandboxes-routes/_helpers.ts
 
 export const sandboxesRouter = new OpenAPIHono();
 
@@ -2172,419 +2105,8 @@ sandboxesRouter.openapi(
 // NOTE: /sandboxes/{id}/run-scripts moved to sandboxes-routes/config.route.ts
 // NOTE: /sandboxes/{id}/stop moved to sandboxes-routes/lifecycle.route.ts
 // NOTE: /sandboxes/{id}/status moved to sandboxes-routes/lifecycle.route.ts
-
-// Publish devcontainer forwarded ports (read devcontainer.json inside instance, expose, persist to Convex)
-sandboxesRouter.openapi(
-  createRoute({
-    method: "post" as const,
-    path: "/sandboxes/{id}/publish-devcontainer",
-    tags: ["Sandboxes"],
-    summary:
-      "Expose forwarded ports from devcontainer.json and persist networking info",
-    request: {
-      params: z.object({ id: z.string() }),
-      body: {
-        content: {
-          "application/json": {
-            schema: z.object({
-              teamSlugOrId: z.string(),
-              taskRunId: z.string(),
-            }),
-          },
-        },
-        required: true,
-      },
-    },
-    responses: {
-      200: {
-        content: {
-          "application/json": {
-            schema: z.array(
-              z.object({
-                status: z.enum(["running"]).default("running"),
-                port: z.number(),
-                url: z.string(),
-              }),
-            ),
-          },
-        },
-        description: "Exposed ports list",
-      },
-      401: { description: "Unauthorized" },
-      500: { description: "Failed to publish devcontainer networking" },
-    },
-  }),
-  async (c) => {
-    const token = await getAccessTokenFromRequest(c.req.raw);
-    if (!token) return c.text("Unauthorized", 401);
-    const { id } = c.req.valid("param");
-    const { teamSlugOrId, taskRunId } = c.req.valid("json");
-    try {
-      // Get instance via provider dispatch
-      const instance = await getInstanceById(id, getMorphClientOrNull());
-
-      const isPveLxc = isPveLxcInstanceId(id);
-
-      const reservedPorts = RESERVED_CMUX_PORT_SET;
-
-      // Attempt to read devcontainer.json for declared forwarded ports
-      const devcontainerJson = await instance.exec(
-        "cat /root/workspace/.devcontainer/devcontainer.json",
-      );
-      const parsed =
-        devcontainerJson.exit_code === 0
-          ? (JSON.parse(devcontainerJson.stdout || "{}") as {
-            forwardPorts?: number[];
-          })
-          : { forwardPorts: [] as number[] };
-
-      const devcontainerPorts = Array.isArray(parsed.forwardPorts)
-        ? (parsed.forwardPorts as number[])
-        : [];
-
-      // Get environmentId from the taskRun (PVE-LXC doesn't persist metadata on instances)
-      const convex = getConvex({ accessToken: token });
-      let environmentPorts: number[] | undefined;
-
-      // First try to get environmentId from the taskRun
-      let environmentId: string | undefined;
-      try {
-        const taskRun = await convex.query(api.taskRuns.get, {
-          teamSlugOrId,
-          id: taskRunId as unknown as string & { __tableName: "taskRuns" },
-        });
-        environmentId = taskRun?.environmentId;
-      } catch {
-        // ignore lookup errors
-      }
-
-      // If we have an environmentId, fetch the environment's exposedPorts
-      if (environmentId) {
-        try {
-          const envDoc = await convex.query(api.environments.get, {
-            teamSlugOrId,
-            id: environmentId as string & {
-              __tableName: "environments";
-            },
-          });
-          environmentPorts = envDoc?.exposedPorts ?? undefined;
-        } catch {
-          // ignore lookup errors; fall back to devcontainer ports
-        }
-      }
-
-      // Build the set of ports we want to expose and persist
-      const allowedPorts = new Set<number>();
-      const addAllowed = (p: number) => {
-        if (!Number.isFinite(p)) return;
-        const pn = Math.floor(p);
-        if (pn > 0 && !reservedPorts.has(pn)) allowedPorts.add(pn);
-      };
-
-      // Prefer environment.exposedPorts if available; otherwise use devcontainer forwardPorts
-      (environmentPorts && environmentPorts.length > 0
-        ? environmentPorts
-        : devcontainerPorts
-      ).forEach(addAllowed);
-
-      const desiredPorts = Array.from(allowedPorts.values()).sort(
-        (a, b) => a - b,
-      );
-      const serviceNameForPort = (port: number) => `port-${port}`;
-
-      let workingInstance = instance;
-      const reloadInstance = async () => {
-        workingInstance = await getInstanceById(instance.id, getMorphClientOrNull());
-      };
-
-      await reloadInstance();
-
-      for (const service of workingInstance.networking.httpServices) {
-        if (!service.name.startsWith("port-")) {
-          continue;
-        }
-        if (reservedPorts.has(service.port)) {
-          continue;
-        }
-        if (!allowedPorts.has(service.port)) {
-          await workingInstance.hideHttpService(service.name);
-        }
-      }
-
-      await reloadInstance();
-
-      for (const port of desiredPorts) {
-        const serviceName = serviceNameForPort(port);
-        const alreadyExposed = workingInstance.networking.httpServices.some(
-          (service) => service.name === serviceName,
-        );
-        if (alreadyExposed) {
-          continue;
-        }
-        try {
-          await workingInstance.exposeHttpService(serviceName, port);
-        } catch (error) {
-          console.error(
-            `[sandboxes.publishNetworking] Failed to expose ${serviceName}`,
-            error,
-          );
-        }
-      }
-
-      // For Morph, reload to get persisted state from their API
-      // For PVE-LXC, skip reload as exposeHttpService only updates in-memory state
-      // and reloading would wipe out the services we just added
-      if (!isPveLxc) {
-        await reloadInstance();
-      }
-
-      const networking = workingInstance.networking.httpServices
-        .filter((s) => allowedPorts.has(s.port))
-        .map((s) => ({ status: "running" as const, port: s.port, url: s.url }));
-
-      // Persist to Convex
-      await convex.mutation(api.taskRuns.updateNetworking, {
-        teamSlugOrId,
-        id: taskRunId as unknown as string & { __tableName: "taskRuns" },
-        networking,
-      });
-
-      return c.json(networking);
-    } catch (error) {
-      console.error("Failed to publish devcontainer networking:", error);
-      return c.text("Failed to publish devcontainer networking", 500);
-    }
-  },
-);
-
-// SSH connection info response schema
-const SandboxSshResponse = z
-  .object({
-    morphInstanceId: z.string(),
-    sshCommand: z.string().describe("Full SSH command to connect to this sandbox"),
-    accessToken: z.string().describe("SSH access token for this sandbox"),
-    user: z.string(),
-    status: z.enum(["running", "paused"]).describe("Current instance status"),
-  })
-  .openapi("SandboxSshResponse");
-
-// Get SSH connection details for a sandbox
-sandboxesRouter.openapi(
-  createRoute({
-    method: "get" as const,
-    path: "/sandboxes/{id}/ssh",
-    tags: ["Sandboxes"],
-    summary: "Get SSH connection details for a sandbox",
-    description:
-      "Returns SSH connection info for a sandbox. Use the returned sshCommand or accessToken to connect.",
-    request: {
-      params: z.object({ id: z.string() }),
-      query: z.object({
-        teamSlugOrId: z.string().optional(),
-      }),
-    },
-    responses: {
-      200: {
-        content: {
-          "application/json": {
-            schema: SandboxSshResponse,
-          },
-        },
-        description: "SSH connection details",
-      },
-      401: { description: "Unauthorized" },
-      403: { description: "Forbidden - not a team member" },
-      404: { description: "Sandbox not found" },
-      500: { description: "Failed to get SSH info" },
-    },
-  }),
-  async (c) => {
-    const user = await getUserFromRequest(c.req.raw);
-    if (!user) {
-      return c.text("Unauthorized", 401);
-    }
-    const { accessToken } = await user.getAuthJson();
-    if (!accessToken) {
-      return c.text("Unauthorized", 401);
-    }
-
-    const { id } = c.req.valid("param");
-    const { teamSlugOrId } = c.req.valid("query");
-
-    try {
-      const convex = getConvex({ accessToken });
-
-      let morphInstanceId: string | null = null;
-
-      // Check if the id is a Morph instance ID (starts with "morphvm_")
-      if (id.startsWith("morphvm_")) {
-        // Direct Morph instance ID - verify ownership via instance metadata
-        const morphClient = getMorphClient();
-
-        // First try to find in task runs if team is provided
-        if (teamSlugOrId) {
-          let taskRun = null;
-          try {
-            taskRun = await convex.query(api.taskRuns.getByContainerName, {
-              teamSlugOrId,
-              containerName: id,
-            });
-          } catch (convexError) {
-            console.log(
-              `[sandboxes.ssh] Convex query failed for ${id}:`,
-              convexError,
-            );
-          }
-
-          if (taskRun) {
-            // Found in task runs - verify team access and that it's a Morph instance
-            await verifyTeamAccess({
-              req: c.req.raw,
-              teamSlugOrId,
-            });
-            if (taskRun.vscode?.provider !== "morph") {
-              return c.text("Sandbox type not supported for SSH", 404);
-            }
-            morphInstanceId = id;
-          }
-        }
-
-        // If not found via task run, verify ownership via instance metadata
-        if (!morphInstanceId) {
-          const result = await verifyInstanceOwnership(
-            morphClient,
-            id,
-            user.id,
-            async () => {
-              const memberships = await convex.query(api.teams.listTeamMemberships, {});
-              return memberships.map((m) => ({ teamId: m.team.teamId }));
-            }
-          );
-          if (!result.authorized) {
-            return c.text(result.message, result.status);
-          }
-          morphInstanceId = result.instanceId;
-        }
-      } else {
-        // For task-run IDs, team is required to look up the task run
-        if (!teamSlugOrId) {
-          return c.text("teamSlugOrId is required for task-run IDs", 400);
-        }
-
-        // Verify team access
-        const team = await verifyTeamAccess({
-          req: c.req.raw,
-          teamSlugOrId,
-        });
-        // Assume it's a task-run ID - look up the sandbox
-        let taskRun: Doc<"taskRuns"> | null = null;
-
-        try {
-          taskRun = await convex.query(api.taskRuns.get, {
-            teamSlugOrId,
-            id: id as Id<"taskRuns">,
-          });
-        } catch {
-          // Not a valid task run ID
-          return c.text("Invalid sandbox or task-run ID", 404);
-        }
-
-        if (!taskRun) {
-          return c.text("Task run not found", 404);
-        }
-
-        // Verify the task run is in the correct team
-        if (taskRun.teamId !== team.uuid) {
-          return c.text("Forbidden", 403);
-        }
-
-        // Check if this task run has an active Morph sandbox
-        if (!taskRun.vscode) {
-          return c.text("No sandbox associated with this task run", 404);
-        }
-
-        if (taskRun.vscode.provider !== "morph") {
-          return c.text("Sandbox type not supported for SSH", 404);
-        }
-
-        if (!taskRun.vscode.containerName) {
-          return c.text("Sandbox container name not found", 404);
-        }
-
-        // Only return SSH info for running/starting sandboxes
-        if (
-          taskRun.vscode.status !== "running" &&
-          taskRun.vscode.status !== "starting"
-        ) {
-          return c.text("Sandbox is not running", 404);
-        }
-
-        morphInstanceId = taskRun.vscode.containerName;
-      }
-
-      if (!morphInstanceId) {
-        return c.text("Could not resolve sandbox instance", 404);
-      }
-
-      // Get SSH access token from Morph API
-      const sshKeyResponse = await fetch(
-        `https://cloud.morph.so/api/instance/${morphInstanceId}/ssh/key`,
-        {
-          method: "GET",
-          headers: {
-            Authorization: `Bearer ${env.MORPH_API_KEY}`,
-            "Content-Type": "application/json",
-          },
-        }
-      );
-
-      if (!sshKeyResponse.ok) {
-        const errorText = await sshKeyResponse.text();
-        console.error(
-          `[sandboxes.ssh] Morph API returned ${sshKeyResponse.status}: ${errorText}`
-        );
-        // Return 404 if the instance doesn't exist in Morph
-        if (sshKeyResponse.status === 404 || errorText.includes("not found")) {
-          return c.text("Sandbox not found", 404);
-        }
-        return c.text("Failed to get SSH credentials", 500);
-      }
-
-      const sshKeyData = (await sshKeyResponse.json()) as {
-        private_key: string;
-        public_key: string;
-        password: string;
-        access_token: string;
-      };
-
-      if (!sshKeyData.access_token) {
-        console.error("[sandboxes.ssh] Morph API did not return access_token");
-        return c.text("Failed to get SSH credentials", 500);
-      }
-
-      // Get instance status from Morph
-      const morphClient = getMorphClient();
-      const instance = await morphClient.instances.get({ instanceId: morphInstanceId });
-      const status = instance.status === "paused" ? "paused" : "running";
-
-      const sshCommand = `ssh ${sshKeyData.access_token}@ssh.cloud.morph.so`;
-      return c.json({
-        morphInstanceId,
-        sshCommand,
-        accessToken: sshKeyData.access_token,
-        user: "root",
-        status,
-      });
-    } catch (error) {
-      if (error instanceof HTTPException) {
-        return c.text(error.message || "Request failed", error.status);
-      }
-      console.error("[sandboxes.ssh] Failed to get SSH info:", error);
-      return c.text("Failed to get SSH info", 500);
-    }
-  },
-);
-
+// NOTE: /sandboxes/{id}/publish-devcontainer moved to sandboxes-routes/features.route.ts
+// NOTE: /sandboxes/{id}/ssh moved to sandboxes-routes/features.route.ts
 // NOTE: /sandboxes/{id}/resume moved to sandboxes-routes/lifecycle.route.ts
 // NOTE: /sandboxes/{id}/discover-repos moved to sandboxes-routes/features.route.ts
 // NOTE: /sandboxes/{id}/live-diff moved to sandboxes-routes/features.route.ts
